test: split page builder drag and resize tests
